Major refactoring of the Input plugin – extracted all the codee to separate utils.
